requirements.txt
# --- Core app ---
Flask>=3.0,<3.1
Flask-Login>=0.6,<0.7
Flask-WTF>=1.2,<1.3
flask-cors>=4.0,<4.1
python-dotenv>=1.0,<1.1
pymongo>=4.6,<5.0
pillow>=10.3,<11.0
jinja2>=3.1,<4.0
werkzeug>=3.0,<4.0
WTForms>=3.1,<4.0

# --- Torch: blocca sulle versioni che hai già installato ---
torch==2.4.1
torchvision==0.19.1

# (Niente torchaudio: non serve e crea conflitti)
